refactor: déplacer plays vers features/plays/
